feature test

+++ src/main/java/dev/iancmd/operationrelife/blocks/DragonEgg.java
@@ -6,30 +6,23 @@ import net.minecraft.util.Mth;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.DragonEggBlock;
-import net.minecraft.world.level.block.Fallable;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.block.state.StateDefinition;
-import net.minecraft.world.level.block.state.properties.BooleanProperty;
-import net.minecraft.world.level.block.state.properties.Property;
 import net.minecraft.world.level.border.WorldBorder;
 import net.minecraft.world.level.material.Material;
 import net.minecraft.world.level.material.MaterialColor;
-import net.minecraft.world.level.pathfinder.PathComputationType;
 import net.minecraft.world.phys.BlockHitResult;
-import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
 
-import static dev.iancmd.operationrelife.OperationRelife.TEST_DIM_KEY;
 import static dev.iancmd.operationrelife.OperationRelife.logger;
-import static org.apache.logging.log4j.Level.*;
+import static org.apache.logging.log4j.Level.DEBUG;
 
 public class DragonEgg extends DragonEggBlock {
   protected static final VoxelShape SHAPE = Block.box(1.0D, 0.0D, 1.0D, 15.0D, 16.0D, 15.0D);
+
   public DragonEgg() {
     super(BlockBehaviour.Properties.of(Material.EGG, MaterialColor.COLOR_BLACK).strength(3.0F, 9.0F).lightLevel((p_50840_) -> {
       return 1;
@@ -38,7 +31,7 @@ public class DragonEgg extends DragonEggBlock {
 
   @Override
   public InteractionResult use(BlockState p_52923_, Level p_52924_, BlockPos p_52925_, Player p_52926_, InteractionHand p_52927_, BlockHitResult p_52928_) {
-    
+
     this.teleport(p_52923_, p_52924_, p_52925_);
     return InteractionResult.sidedSuccess(p_52924_.isClientSide);
   }
@@ -49,7 +42,7 @@ public class DragonEgg extends DragonEggBlock {
   }
 
   private void teleport(BlockState p_52936_, Level p_52937_, BlockPos p_52938_) {
-    logger.log(DEBUG,p_52937_.getBlockState(p_52938_));
+    logger.log(DEBUG, p_52937_.getBlockState(p_52938_));
     WorldBorder worldborder = p_52937_.getWorldBorder();
     for (int i = 0; i < 1000; ++i) {
       BlockPos blockpos = p_52938_.offset(p_52937_.random.nextInt(16) - p_52937_.random.nextInt(16), p_52937_.random.nextInt(8) - p_52937_.random.nextInt(8), p_52937_.random.nextInt(16) - p_52937_.random.nextInt(16));
@@ -60,10 +53,10 @@ public class DragonEgg extends DragonEggBlock {
             float f = (p_52937_.random.nextFloat() - 0.5F) * 0.2F;
             float f1 = (p_52937_.random.nextFloat() - 0.5F) * 0.2F;
             float f2 = (p_52937_.random.nextFloat() - 0.5F) * 0.2F;
-            double d1 = Mth.lerp(d0, (double) blockpos.getX(), (double) p_52938_.getX()) + (p_52937_.random.nextDouble() - 0.5D) + 0.5D;
-            double d2 = Mth.lerp(d0, (double) blockpos.getY(), (double) p_52938_.getY()) + p_52937_.random.nextDouble() - 0.5D;
-            double d3 = Mth.lerp(d0, (double) blockpos.getZ(), (double) p_52938_.getZ()) + (p_52937_.random.nextDouble() - 0.5D) + 0.5D;
-            p_52937_.addParticle(ParticleTypes.PORTAL, d1, d2, d3, (double) f, (double) f1, (double) f2);
+            double d1 = Mth.lerp(d0, blockpos.getX(), p_52938_.getX()) + (p_52937_.random.nextDouble() - 0.5D) + 0.5D;
+            double d2 = Mth.lerp(d0, blockpos.getY(), p_52938_.getY()) + p_52937_.random.nextDouble() - 0.5D;
+            double d3 = Mth.lerp(d0, blockpos.getZ(), p_52938_.getZ()) + (p_52937_.random.nextDouble() - 0.5D) + 0.5D;
+            p_52937_.addParticle(ParticleTypes.PORTAL, d1, d2, d3, f, f1, f2);
           }
         } else {
           p_52937_.setBlock(blockpos, p_52936_, 2);
